requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0
google-genai>=0.3.0
